[Docs] 상품 엔티티 및 DTO 치수 타입 수정 상/하의 치수 타입 Integer -> Float 수정

+++ src/main/java/com/example/lonua/product/model/entity/Product.java
@@ -69,17 +69,17 @@ public class Product {
     @Column(nullable = false)
     private Integer price;
 
-    private Integer shoulderWidth; // 어깨 너비
-    private Integer chestSize;  // 가슴 둘레
-    private Integer armLength;  // 팔 길이
-    private Integer topLength;  // 상의 총 길이
+    private Float shoulderWidth; // 어깨 너비
+    private Float chestSize;  // 가슴 둘레
+    private Float armLength;  // 팔 길이
+    private Float topLength;  // 상의 총 길이
 
-    private Integer waistline;  // 허리 둘레
-    private Integer hipCircumference;  // 엉덩이 둘레
-    private Integer thighCircumference;  // 허벅지 둘레
-    private Integer crotchLength;  // 밑위 길이
-    private Integer hemLength;  // 밑단 길이
-    private Integer totalBottomLength;  // 하의 총 길이
+    private Float waistline;  // 허리 둘레
+    private Float hipCircumference;  // 엉덩이 둘레
+    private Float thighCircumference;  // 허벅지 둘레
+    private Float crotchLength;  // 밑위 길이
+    private Float hemLength;  // 밑단 길이
+    private Float totalBottomLength;  // 하의 총 길이
 
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
